:recycle: refactor: 이메일 알림 생성용 레포지토리 리팩토링
- projection 대신 생성자 사용

- transform 메서드를 사용하여 Map 타입으로 데이터 조회

- 불필요한 DTO 삭제
